Move subprocess helpers from `update-release-branch.ts` to `command.ts`
